pipeline_analise_sentimento_mercado/pyproject.toml
[project]
name = "pipeline-analise-sentimento-mercado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]
